refactor infolojo.js and extract logic to two new files scrollHelper.js and infolojoLogger.js
